navigationStore caching problem
If the navigation is modified, if the APP obtains the navigationStore from the sotre and finds that the route does not exist, it will report an error. I think this mode is very good for the start mode. It can be refreshed to stay in the current interface, but the released version, if this mode is also enabled. Even if you cover a new application, there will be this problem.