Prettier-ize
Don't bother prettying ignite.json:

Ignite (indirectly thru gluegun) uses fs-jetback to write this file;
there's no easy way to tell it to make sure to write a newline on the
end, and this file is written after Bowser finishes so we can't fix it.

Rather than leave this as a difference that'll be created the first
time someone runs `yarn format` in the new project, we'll ignore it.