Convert class components to functions
I converted StorybookUIRoot to a function (in spite of the comment):

- I removed its "RN hot module must be in a class for HMR" comment; I
tested this and found no problems with hot reloading enabled, for this
component or the other components I've converted to functions.

- Googling led me to https://mobx.js.org/best/stateless-HMR.html, which
I think might also be old information; I tested the newly-functional
components with react-devtools and it's displaying their names, so I
think something must've changed with transpiling to fix that problem.

Also, the new functional components seem to trigger warnings from
the react/prop-types eslint rule. We don't put propTypes on our
components because we've got Typescript to watch our back. So, I've
turned off this eslint rule.