Don't bump mobx for now; stay at 4.x.x to keep iOS 9 & Hermes compatibility
We need to lock mobx at 4.x because of facebook/hermes#28 (as well as the
old version of JavaScriptCore in iOS 9 not supporting Proxies). We could
move to 5.x if we are okay with not supporting iOS 9 and Hermes, but we
don't think that's a viable path forward at this time.